move fetchData to app component
