+++ src/App.jsx
@@ -1,8 +1,24 @@
 import './App.css';
 import Filters from './Filters';
 import { getResources } from './config/firebase-functions';
+import { useState, useEffect } from 'react'
 
-function App({data}) {
+function App() {
+  const [data, setData] = useState([]);
+
+  //Fetch data from Firestore database and set to state
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const dataFromFirestore = await getResources();
+        setData(dataFromFirestore || []);
+      } catch (error) {
+        console.error('Error fetching data from Firestore:', error);
+      }
+    };
+
+    fetchData();
+  }, []);
 
 
   return (
@@ -21,7 +37,7 @@ function App({data}) {
                 </div>
             </section>
 
-            <Filters />
+            <Filters data={data}/>
               
     </>
   )
